\name{ENMTools-package}
\alias{ENMTools-package}
\alias{ENMTools}
\docType{package}
\title{
ENMTools
}
\description{
At present ENMTools R package contains functions for performing tests from Warren et al. 2008, but it will eventually contain most or all of the functions from the ENMTools Perl script/executable file available from enmtools.com.
}
\details{
\tabular{ll}{
Package: \tab ENMTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-11-27\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
